EOVERFLOW if output string is larger than INT_MAX
If the number of bytes required to hold the output string (excluding the
nul-termination) is larger than INT_MAX, return -1 and set errno to
EOVERFLOW (if available).  Before, we used ERANGE in this case and we
also did that if the number of required bytes was equal to INT_MAX.  The
latter behaviour had the advantage that users could safely increment any
positive return value by one (in order to calculate the required buffer
size) even without converting it to a larger or unsigned integer type.
However, as other implementations may return INT_MAX, users should
handle this case anyway.  The new behaviour is defined by an XSI
extension to POSIX (IEEE Std 1003.1, 2004 Edition), though this
extension also mandates to fail with EOVERFLOW if the buffer size
specified by the user is greater than INT_MAX; which we won't do, as
this alone is not a problem whatsoever.  Note that C99 doesn't specify
the error handling.